# source/public/docs/api-docs.yaml
swagger: "2.0"
info:
  description: "This is todock API docs. Copyrighted by code-idea."
  version: "1.0.0"
  title: "Swagger Todock (토닥) - 작성중..."
host: "todak.codeidea.io"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /partners/one:
    get:
      tags:
      - "거래처 (병원)"
      summary: "거래처를 조회한다."
      operationId: "partner"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - id: "1"
        in: "query"
        required: true
        type: "integer"
        maximum: 9999.0
        minimum: 1.0
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResult"
...
definitions:
...
  ApiResult:
    type: "object"
    properties:
      data:
        type: "object"
      ment:
        type: "string"
      result:
        type: "boolean"
...